chore: rename filter-routes to controller/routes
